+++ Source/Project64-core/N64System/Recompiler/x86/x86RecompilerOps.cpp
@@ -7362,12 +7362,12 @@ void x86_compiler_COP0_CO_ERET()
 {
     if ((g_Reg->STATUS_REGISTER & STATUS_ERL) != 0)
     {
-        g_Reg->m_PROGRAM_COUNTER = g_Reg->ERROREPC_REGISTER;
+        g_Reg->m_PROGRAM_COUNTER = (uint32_t)g_Reg->ERROREPC_REGISTER;
         g_Reg->STATUS_REGISTER &= ~STATUS_ERL;
     }
     else
     {
-        g_Reg->m_PROGRAM_COUNTER = g_Reg->EPC_REGISTER;
+        g_Reg->m_PROGRAM_COUNTER = (uint32_t)g_Reg->EPC_REGISTER;
         g_Reg->STATUS_REGISTER &= ~STATUS_EXL;
     }
     g_Reg->m_LLBit = 0;
